# src/api/docs/swagger.yaml
basePath: /
definitions:
  handler.APIResponse:
    properties:
      message:
        description: need fill only if error occurred
        type: string
    type: object
...
  handler.IngredientAPIResponse:
    properties:
      item:
        $ref: '#/definitions/ingredient.Ingredient'
        type: object
      message:
        description: need fill only if error occurred
        type: string
    type: object
...
  handler.ListIngredientsAPIResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/ingredient.Ingredient'
        type: array
      message:
        description: need fill only if error occurred
        type: string
    type: object
...
  ingredient.Ingredient:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  services.CreateIngredientRequest:
    properties:
      name:
        description: (required)
        maxLength: 255
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  services.UpdateIngredientRequest:
    properties:
      name:
        description: (required)
        maxLength: 255
        minLength: 3
        type: string
    required:
    - name
    type: object
...
host: api.food.test
info:
  contact: {}
  description: This is Swagger docs for Food API
  license: {}
  title: Food API
  version: "1.0"
paths:
...
  /v1/ingredients/:
    get:
      description: find ingredients by params
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListIngredientsAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get ingredients
      tags:
      - receipts
    post:
      parameters:
      - description: params
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/services.CreateIngredientRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.IngredientAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create ingredient
      tags:
      - receipts
  /v1/ingredients/{id}:
    put:
      parameters:
      - description: Ingredient id
        in: path
        name: id
        required: true
        type: integer
      - description: params
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/services.UpdateIngredientRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.IngredientAPIResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update ingredient
      tags:
      - receipts
  /v1/media/{folder}/{filename}:
    get:
      description: get media
      parameters:
      - description: 1ac4f9a135d204e71ed41fa8accfbe42
        in: path
        name: folder
        required: true
        type: string
      - description: avatar.png
        in: path
        name: filename
        required: true
        type: string
      produces:
      - image/png
      - image/jpeg
      - image/gif
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get media from DB
      tags:
      - media
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"